editor/libeditor/html/tests/test_bug629845.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=629845
-->
<head>
  <title>Test for Bug 629845</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=629845">Mozilla Bug 629845</a>
<p id="display"></p>

<script>
function initFrame(frame)
{
  frame.contentWindow.document.designMode="on";
  frame.contentWindow.document.writeln("<body></body>");

  document.getElementsByTagName('button')[0].click();
}

function command(aName)
{
  var frame = document.getElementsByTagName('iframe')[0];

  is(frame.contentDocument.designMode, "on", "design mode should be on!");
  var caught = false;
  try {
    frame.contentDocument.execCommand(aName, false, null);
  } catch (e) {
    ok(false, "exception " + e + " was thrown");
    caught = true;
  }

  ok(!caught, "No exception should have been thrown.");

  // Stop the document load before finishing, just to be clean.
  document.getElementsByTagName('iframe')[0].contentWindow.document.close();
  SimpleTest.finish();
}
</script>

<div id="content">
  <button type="button" onclick="command('bold');">Bold</button>
  <iframe onload="initFrame(this);"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 629845 **/

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
